Run graph Asynchronously/pyproject.toml
[project]
name = "run-graph-asynchronously"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langgraph>=0.5.2",
    "python-dotenv>=1.1.1",
]
